refactor(agent): relocate runtime leaves (types + failure-tracker)
Bottom-up step toward moving AgentRuntime out of src/legacy/. The two
leaf modules — runtime types and failure-tracker — move first into a
new src/agent/runtime/ folder. runtime.ts itself stays put for now and
imports them from the new location.
