+++ src/agent/runtime/failure-tracker.ts
